// src/selectors/aws-credential/models/IAWSCredentialTableRow.ts
export default interface IAWSCredentialTableRow {
    readonly awsid: string;
    readonly accessKey: string;
    readonly secretKey: string;
    readonly notes: string;
    readonly createdBy: string;
    readonly createdDate: string;
    readonly createDate: string;
    readonly modifiedBy: string;
    readonly modifiedDate: string;
    readonly createUserID: number;
  }
  